refactor: move search bar to its own component
